+++ DoPrebuild.props
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
-    <ExecutePrebuild>false</ExecutePrebuild>
+    <ExecutePrebuild>true</ExecutePrebuild>
   </PropertyGroup>
 </Project>
