+++ JotunnModStub.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31019.35
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JotunnModStub", "JotunnModStub\JotunnModStub.csproj", "{DEAF4438-8089-40ED-8175-398E1261D45B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spawners", "JotunnModStub\Spawners.csproj", "{DEAF4438-8089-40ED-8175-398E1261D45B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
